+++ autd3-driver/src/datagram/reads_fpga_info.rs
@@ -4,7 +4,7 @@
  * Created Date: 06/12/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 06/12/2023
+ * Last Modified: 30/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -20,7 +20,7 @@ pub struct ConfigureReadsFPGAInfo<F: Fn(&Device) -> bool> {
 
 impl<F: Fn(&Device) -> bool> ConfigureReadsFPGAInfo<F> {
     /// constructor
-    pub fn new(f: F) -> Self {
+    pub const fn new(f: F) -> Self {
         Self { f }
     }
 }