+++ examples/src/tests/stm.rs
@@ -4,7 +4,7 @@
  * Created Date: 28/05/2021
  * Author: Shun Suzuki
  * -----
- * Last Modified: 14/11/2023
+ * Last Modified: 30/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2021 Shun Suzuki. All rights reserved.
@@ -14,13 +14,13 @@
 use autd3::prelude::*;
 
 pub async fn focus_stm<L: Link>(autd: &mut Controller<L>) -> anyhow::Result<bool> {
-    autd.send(Silencer::disable()).await?;
+    autd.send(ConfigureSilencer::disable()).await?;
 
     let center = autd.geometry.center() + Vector3::new(0., 0., 150.0 * MILLIMETER);
 
     let point_num = 200;
     let radius = 30.0 * MILLIMETER;
-    let stm = FocusSTM::new(1.0).add_foci_from_iter((0..point_num).map(|i| {
+    let stm = FocusSTM::from_freq(1.0).add_foci_from_iter((0..point_num).map(|i| {
         let theta = 2.0 * PI * i as float / point_num as float;
         let p = radius * Vector3::new(theta.cos(), theta.sin(), 0.0);
         center + p
@@ -34,14 +34,14 @@ pub async fn focus_stm<L: Link>(autd: &mut Controller<L>) -> anyhow::Result<bool
 }
 
 pub async fn gain_stm<L: Link>(autd: &mut Controller<L>) -> anyhow::Result<bool> {
-    autd.send(Silencer::disable()).await?;
+    autd.send(ConfigureSilencer::disable()).await?;
 
     let center = autd.geometry.center() + Vector3::new(0., 0., 150.0 * MILLIMETER);
 
     let point_num = 50;
     let radius = 30.0 * MILLIMETER;
 
-    let stm = GainSTM::new(1.0).add_gains_from_iter((0..point_num).map(|i| {
+    let stm = GainSTM::from_freq(1.0).add_gains_from_iter((0..point_num).map(|i| {
         let theta = 2.0 * PI * i as float / point_num as float;
         let p = radius * Vector3::new(theta.cos(), theta.sin(), 0.0);
         Focus::new(center + p)