+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# 19.1.1
+# 20.0.0
 
+- Update firmware to v5.0.0
+  - Add fixed completion steps algorithm of Silencer  
 - Fix `T4010A1_AMPLITUDE` value
+- Add calculation of index of Modulation and STM from system time in firmware-emulator
 
 # 19.1.0